Complete _store coverage by recording tests against the _store entry points.
