+++ snapcraft/tests/store_tests/test_download.py
@@ -17,10 +17,7 @@ import os
 
 import testscenarios
 
-from snapcraft import (
-    _store,
-    storeapi,
-)
+from snapcraft import storeapi
 from snapcraft.tests import store_tests
 
 
@@ -118,37 +115,3 @@ Getting details for {pkg_name}
 Downloading {pkg_name}
 Successfully downloaded {pkg_name} at downloaded.snap
 '''.format(pkg_name=pkg_name))
-
-
-class TestDownload_store(store_tests.RecordedTestCase):
-
-    def test_download_without_login(self):
-        pkg_name = 'femto'
-        _store.download(pkg_name, 'stable', 'downloaded.snap', 'amd64')
-        self.assertIn('No valid credentials found', self.logger.output)
-
-    def test_download_not_found(self):
-        self.login()
-        self.addCleanup(self.logout)
-        exc = self.assertRaises(
-            RuntimeError,
-            _store.download, 'gloo', 'bee', 'downloaded.snap', 'amd64')
-        self.assertEqual(
-            'Snap gloo for amd64 cannot be found in the bee channel',
-            str(exc))
-
-    def test_download_mismatch(self):
-        self.login()
-        self.addCleanup(self.logout)
-        self.addCleanup(setattr, storeapi.SCAClient, 'download',
-                        storeapi.SCAClient.download)
-
-        def raise_not_a_sha(*args):
-            raise storeapi.SHAMismatchError('downloaded.snap', 'not-a-sha')
-        storeapi.SCAClient.download = raise_not_a_sha
-        exc = self.assertRaises(
-            RuntimeError,
-            _store.download, 'femto', 'stable', 'downloaded.snap', 'amd64')
-        self.assertEqual(
-            'Failed to download femto at downloaded.snap (mismatched SHA)',
-            str(exc))
