+++ snapcraft/tests/store_tests/test_upload.py
@@ -13,17 +13,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import contextlib
 import os
 import unittest
 
 import requests
 import testscenarios
 
-from snapcraft import (
-    _store,
-    storeapi,
-)
+from snapcraft import storeapi
 from snapcraft.storeapi import _upload
 from snapcraft.tests import store_tests
 
@@ -172,122 +168,3 @@ class ScanStatusTestCase(unittest.TestCase):
 
     def test_get_status_http_error(self):
         self.assertIsNone(self.get_scan_status(requests.HTTPError))
-
-
-class TestUpload_store(store_tests.TestCase):
-
-    def test_upload_unknwon_file(self):
-        self.login()
-        self.addCleanup(self.logout)
-        exc = self.assertRaises(
-            FileNotFoundError,
-            _store.upload, 'I-dont-exist')
-        self.assertEqual('I-dont-exist', str(exc))
-
-    def test_upload_fails(self):
-        self.login()
-        self.addCleanup(self.logout)
-        path, name = self.create_snap('basic')
-
-        def you_failed(*args):
-            return dict(success=False, errors=['You failed'])
-        self.addCleanup(
-            setattr, storeapi.SCAClient, 'upload', storeapi.SCAClient.upload)
-        storeapi.SCAClient.upload = you_failed
-        _store.upload(path)
-        self.assertIn('Upload did not complete', self.logger.output)
-        self.assertIn('Some errors were detected:\n\nYou failed\n',
-                      self.logger.output)
-
-    def test__upload_files_fails_review(self):
-        # Pretend the binary upload succeeded
-        def uploaded(name, data):
-            class Response(object):
-                ok = True
-
-                def json(*args):
-                    return dict(status_url='foobar')
-            return Response()
-        self.store.upload_snap = uploaded
-
-        # Pretend the snap upload failed the scan
-        class FailingExecutor(object):
-
-            def done(self):
-                return True
-
-            def result(self):
-                return True, dict(message='You failed review')
-
-        class Spawner(object):
-
-            def submit(*args):
-                return FailingExecutor()
-
-        @contextlib.contextmanager
-        def executor(*args, **kwargs):
-            yield Spawner()
-
-        self.addCleanup(
-            setattr, _upload, 'ThreadPoolExecutor', _upload.ThreadPoolExecutor)
-        _upload.ThreadPoolExecutor = executor
-        result = {}
-        res = _upload._upload_files(self.store, 'foo', {}, result)
-        self.assertEqual(['You failed review'], res['errors'])
-
-    def test__upload_files_review_too_long(self):
-        # Pretend the binary upload succeeded
-        def uploaded(name, data):
-            class Response(object):
-                ok = True
-
-                def json(*args):
-                    return dict(status_url='foobar',
-                                web_status_url='See there')
-            return Response()
-        self.store.upload_snap = uploaded
-
-        # Pretend the snap upload failed the scan
-        class FailingExecutor(object):
-
-            def done(self):
-                return True
-
-            def result(self):
-                return False, dict(message='You failed review')
-
-        class Spawner(object):
-
-            def submit(*args):
-                return FailingExecutor()
-
-        @contextlib.contextmanager
-        def executor(*args, **kwargs):
-            yield Spawner()
-
-        self.addCleanup(
-            setattr, _upload, 'ThreadPoolExecutor', _upload.ThreadPoolExecutor)
-        _upload.ThreadPoolExecutor = executor
-        result = {}
-        res = _upload._upload_files(self.store, 'foo', {}, result)
-        self.assertEqual(['Package scan took too long.',
-                          'Please check the status later at: See there.'],
-                         res['errors'])
-
-    def test__upload_files_other_error(self):
-        # Pretend the binary failed to upload
-        def uploaded(name, data):
-            class Response(object):
-                ok = False
-                reason = 'Some'
-                text = 'Full Moon'
-            return Response()
-        self.store.upload_snap = uploaded
-        result = {}
-        res = _upload._upload_files(self.store, 'foo', {}, result)
-        self.assertEqual(['Full Moon'],
-                         res['errors'])
-        self.assertIn('There was an error uploading the application.\n'
-                      'Reason: Some\n'
-                      'Text: Full Moon',
-                      self.logger.output)
